Fix leaderboard viewer
