+++ lux-eye-s2/src/pages/home/HomePage.tsx
@@ -1,5 +1,6 @@
 import { Container, Stack, Text } from '@mantine/core';
-import { Navigate } from 'react-router-dom';
+import { useCallback, useEffect } from 'react';
+import { Navigate, useNavigate } from 'react-router-dom';
 import { useStore } from '../../store';
 import { notifyError } from '../../utils/notifications';
 import { HomeCard } from './HomeCard';
@@ -13,9 +14,38 @@ declare global {
   }
 }
 
+let hasData = false;
+
 export function HomePage(): JSX.Element {
   const episode = useStore(state => state.episode);
   const load = useStore(state => state.load);
+
+  const navigate = useNavigate();
+
+  const onWindowMessage = useCallback((event: MessageEvent<any>) => {
+    if (hasData) {
+      return;
+    }
+
+    if (event.data && event.data.environment) {
+      hasData = true;
+
+      try {
+        load(event.data.environment);
+        navigate('/leaderboard');
+      } catch (err: any) {
+        console.error(err);
+        notifyError('Cannot load episode from Kaggle', err.message);
+      }
+    }
+  }, []);
+
+  useEffect(() => {
+    window.addEventListener('message', onWindowMessage);
+    return () => {
+      window.removeEventListener('message', onWindowMessage);
+    };
+  });
 
   if (episode === null && window.episode !== undefined) {
     try {
